drivers/tcu: handle TCU messages asynchronously.
That is, we now no longer poll until all sidecalls arrive in a specific
order, but use an interrupt handler to handle sidecalls. To make that
work, I've also restructured the code to store all state in the
tcu_device struct and pass it to basically all functions. Additionally,
we now use platform_driver to get the IRQ from the DTS.

There are still some TODOs to be addressed, but it seems to work fine
(on hw).