drivers/tcu: support polling/blocking on our TCU device.
We wake the blocked task up again if we get a foreign-message IRQ from
the TCU. Additionally, we need to do that when noticing pending
messages for the old task after a context switch. This is rather
tricky, because we cannot wake up tasks during the context switch.
Thus, we do that later before returning from exceptions.