drivers/tcu: insert TLB entries with the permissions from the page.
If we only insert the requested permission, we might reinsert the same
page all the time with different permissions. For example, if the same
buffer is used for reading and writing, we will alternate between a
page with read and write permission in the TLB.

Therefore, we now lookup the permission in the process's page table and
use this permission. We also check whether the requested permissions
are present and error if not.